Move Limewire main to __init__.py
